backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
black==25.1.0
    # via -r requirements.in
boto3==1.37.22
    # via -r requirements.in
botocore==1.37.22
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
cffi==1.17.1
    # via
    #   -r requirements.in
    #   cryptography
click==8.1.8
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
cryptography==44.0.2
    # via -r requirements.in
django==5.1.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.7.0
    # via -r requirements.in
django-storages==1.14.5
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
flake8==7.2.0
    # via -r requirements.in
isort==6.0.1
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   -r requirements.in
    #   boto3
    #   botocore
markupsafe==3.0.2
    # via -r requirements.in
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
mypy==1.15.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
    #   mypy
packaging==24.2
    # via
    #   -r requirements.in
    #   black
    #   build
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.7
    # via
    #   -r requirements.in
    #   black
psycopg2-binary==2.9.10
    # via -r requirements.in
pycodestyle==2.13.0
    # via
    #   -r requirements.in
    #   flake8
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pyflakes==3.3.2
    # via
    #   -r requirements.in
    #   flake8
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
python-decouple==3.8
    # via -r requirements.in
s3transfer==0.11.4
    # via
    #   -r requirements.in
    #   boto3
six==1.17.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
tomli==2.2.1
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   mypy
tzdata==2025.1
    # via
    #   -r requirements.in
    #   django
urllib3==2.3.0
    # via
    #   -r requirements.in
    #   botocore
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools